Streamline ARFHoeffdingTreeClassifier nodes
